Moved several highlighting related classes to the highlighting package.
