+++ resources/META-INF/plugin.xml
@@ -18,10 +18,10 @@
   <extensions defaultExtensionNs="com.intellij">
     <fileTypeFactory implementation="nl.rubensten.texifyidea.LatexFileTypeFactory"/>
     <lang.parserDefinition language="Latex" implementationClass="nl.rubensten.texifyidea.LatexParserDefinition"/>
-    <lang.syntaxHighlighterFactory language="Latex" implementationClass="nl.rubensten.texifyidea.LatexSyntaxHighlighterFactory"/>
+    <lang.syntaxHighlighterFactory language="Latex" implementationClass="nl.rubensten.texifyidea.highlighting.LatexSyntaxHighlighterFactory"/>
     <lang.commenter language="Latex" implementationClass="nl.rubensten.texifyidea.LatexCommenter"/>
     <lang.braceMatcher language="Latex" implementationClass="nl.rubensten.texifyidea.LatexPairedBraceMatcher"/>
-    <colorSettingsPage implementation="nl.rubensten.texifyidea.LatexColorSettingsPage"/>
+    <colorSettingsPage implementation="nl.rubensten.texifyidea.highlighting.LatexColorSettingsPage"/>
     <completion.contributor language="Latex" implementationClass="nl.rubensten.texifyidea.completion.LatexCompletionContributor"/>
     <codeInsight.lineMarkerProvider language="Latex" implementationClass="nl.rubensten.texifyidea.LatexLineMarkerProvider"/>
