+++ src/nl/rubensten/texifyidea/highlighting/LatexColorSettingsPage.java
@@ -1,10 +1,11 @@
-package nl.rubensten.texifyidea;
+package nl.rubensten.texifyidea.highlighting;
 
 import com.intellij.openapi.editor.colors.TextAttributesKey;
 import com.intellij.openapi.fileTypes.SyntaxHighlighter;
 import com.intellij.openapi.options.colors.AttributesDescriptor;
 import com.intellij.openapi.options.colors.ColorDescriptor;
 import com.intellij.openapi.options.colors.ColorSettingsPage;
+import nl.rubensten.texifyidea.TexifyIcons;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
